Lesson learned: low memory cant be mapped by 4k
apparently th low page tables (with p4 and p3 offsets of 0) are x86_64
huge pages, meaning they are 2MB each. this means for example that
virtual address 0x115000 cannot be mapped to physical address 0x114000.
Since this is what was being used for a test case, much more time than
necessary was spent debugging this issue. For future note, an easy demo
in low memory is mapping 0x314000 to 0x114000. notice that the
difference here is 0x200000, or 2MB, which is exactly the huge page
size!