virtual memory management implemented!
the basic premise of the memory manager is that there is an 8 byte cap
on both ends of any "allocated" block of memory. these caps are a packed
data structure which contains the size of the region and the allocation
status. The total size of the allocated chunk, including the caps on
both ends must be a power of 2 greater than 64. This allows the caps to
use the 5 least significant bits of size to store metadata, such as
allocation state. The memory is then treated like a linked list where
free regions are coalesced and allocated regions are left separate from
their allocated neighbors.

There is code in this commit which is for now not compiled due to the
use of ifdefs. The purpose of this code is to implement an avl of free
regions such that the best fitting region may be selected for a new
allocation rather than the first region which has enough space.

In addition, this changeset includes a fix for a bug where frames
containing page tables were being re-allocated.

Signed-off-by: Ted Meyer <tmathmeyer@gmail.com>